강화 버튼 업데이트
강화 버튼을 일부 만들었습니다.
우측 - 연료
하단 - 공속, 이속
좌측 - 공격력

강화시, 플레이어 스텟이 올라가도록 구현

+++ Project1/Project1/Enforce.h
@@ -26,7 +26,7 @@ public:
 
 	void Set_Pos(double, double); // 위치 지정 함수
 	void Move_Mid_Pos(Point rectViewMid, int intrv); // Set_Pos 기반 중앙 위치 이동 함수
-	void Set_TAP(int, int, int); // type, amount, price 지정 함수
+	void Set_TAP(int, double, int); // type, amount, price 지정 함수
 	void Set_Draw(int); // draw 지정
 	void Set_Drawing(int, int, int); // 강화시 그릴 버튼 지정 함수 최대 3개
 	void Set_Open(int); // open 지정
@@ -36,7 +36,7 @@ public:
 protected:
 	double x, y; // 버튼의 위치 배율 (최초 노드는 0,0)
 	int type; // 강화 할 타입
-	int amount; // type 의 직접적인 수치
+	double amount; // type 의 직접적인 수치
 	int price; // 가격
 
 	int draw; // 현재 이 버튼을 그려낼 것인지
